+++ src/merger/paraver/HardwareCounters.h
@@ -72,7 +72,7 @@ typedef struct _cQueue
 {
   struct _cQueue *next, *prev;
 
-  long long Events[MAX_HWC];
+  int Events[MAX_HWC];
   int Traced[MAX_HWC];          /*
                                  * * Boolean field for each counter that
                                  * *  indicates if counter is readed or not.
@@ -84,14 +84,14 @@ CntQueue;
 extern CntQueue CountersTraced;
 
 int HardwareCounters_Emit (int ptask, int task, int thread,
-  long long time, event_t * Event, unsigned int *outtype,
-  unsigned long long *outvalue);
+  unsigned long long time, event_t * Event,
+	unsigned int *outtype, unsigned long long *outvalue);
 void HardwareCounters_Show (event_t * Event, int ncounters);
 void HardwareCounters_Get (event_t *Event, unsigned long long *buffer);
 void HardwareCounters_NewSetDefinition (int ptask, int task, int thread, int newSet, long long *HWCIds);
 int * HardwareCounters_GetSetIds(int ptask, int task, int thread, int set_id);
 int HardwareCounters_GetCurrentSet(int ptask, int task, int thread);
-void HardwareCounters_Change (int ptask, int task, int thread, int newSet, unsigned int *outtypes, unsigned long long *outvalues);
+void HardwareCounters_Change (int ptask, int task, int thread, int newSet, int *outtypes, unsigned long long *outvalues);
 void HardwareCounters_SetOverflow (int ptask, int task, int thread, event_t *Event);
 
 #if defined(PARALLEL_MERGE)
